python/requirements.txt
Flask==3.1.0
PyYAML==6.0.2
Telethon==1.38.1
pytest==8.3.4
requests==2.32.3
python-dotenv==1.0.1
